Adds new library for ssh connections
This commit adds a new library for ssh connections with a list of
tuples. The first value of the tuple ist the key of the url and the
second one is the url itself.

There could be a list e.g.
sshList = [("server1", "server1.mycomp.com")]

If the ssh-prompt asks you for a server, you can take the key of the
lists of tuples for connecting to the long version of the url.