Adds new Prompt to the library
This will add a new possibility to add a prompt for 'switch'
scripts. This scripts has to be three special arguments 'on', 'off' and
'status'. 'on' turns on the switch, 'off' turns off the switch and
'status' will be used for the notification message. For the notification
twmnd and twmnc will be used.

https://github.com/sboli/twmn